+++ aten/src/ATen/core/SymIntArrayRef.h
@@ -11,10 +11,7 @@
 
 #include <ATen/core/SymInt.h>
 #include <c10/util/ArrayRef.h>
-#include <c10/util/C++17.h>
-#include <c10/util/Deprecated.h>
 #include <c10/util/Exception.h>
-#include <c10/util/SmallVector.h>
 
 #include <array>
 #include <initializer_list>
