+++ aten/src/ATen/templates/DispatchKeyFunction.h
@@ -11,31 +11,9 @@
 // Forward declarations of any types needed in the operator signatures.
 // We can't directly include these classes because it will cause circular include dependencies.
 // This file is included by TensorBody.h, which defines the Tensor class.
-namespace c10 {
-
-template<typename T>
-class optional;
-template<typename T>
-class List;
-class Stream;
-class Scalar;
-struct Storage;
-struct TensorOptions;
-
-}
+#include <ATen/core/ATen_fwd.h>
 
 namespace at {
-
-class Tensor;
-struct Dimname;
-struct Generator;
-using TensorList = c10::ArrayRef<Tensor>;
-using DimnameList = c10::ArrayRef<Dimname>;
-using c10::Stream;
-using c10::Storage;
-using c10::QScheme;
-using c10::Scalar;
-using c10::TensorOptions;
 
 namespace ${dispatch_namespace} {
