+++ aten/src/ATen/templates/Functions.h
@@ -69,6 +69,7 @@
 #include <c10/core/TensorOptions.h>
 #include <c10/util/Deprecated.h>
 #include <c10/util/Optional.h>
+#include <c10/util/OptionalArrayRef.h>
 
 #include <ATen/ops/from_blob.h>
 #include <ATen/ops/tensor.h>
