+++ aten/src/ATen/templates/MethodOperators.h
@@ -13,33 +13,7 @@
 // Forward declarations of any types needed in the operator signatures.
 // We can't directly include these classes because it will cause circular include dependencies.
 // This file is included by TensorBody.h, which defines the Tensor class.
-namespace c10 {
-
-template<typename T>
-class optional;
-template<typename T>
-class List;
-class Stream;
-class Scalar;
-struct Storage;
-struct TensorOptions;
-
-}
-
-namespace at {
-
-class Tensor;
-struct Dimname;
-struct Generator;
-using TensorList = c10::ArrayRef<Tensor>;
-using DimnameList = c10::ArrayRef<Dimname>;
-using c10::Stream;
-using c10::Storage;
-using c10::QScheme;
-using c10::Scalar;
-using c10::TensorOptions;
-
-}
+#include <ATen/core/ATen_fwd.h>
 
 ${MethodOperators_includes}
